R file adding the distance to city control to data
This file uses the "cycling density for did" and "cycling access for did" files to add a control to the dataset: the distance from each centroid to the city's center. This file creates two final datasets: "cycling_density_dist.RDS" and "paris1421_cyclingaccess_dist.RDS", which now have their respective measurements of bike lane density or distance to the nearest bike lane but also both include the distance to city center control which is used in the final regression testing.