preliminary code defining function
preliminary code defining a new function called "project to cycling" which aims to project the centroid's coordinates to the nearest cycling lane, thus creating a framework to calculate the distance from the centroid's center to the nearest bike lane. This function was created, though not finished, by Dr. Jonathan Eastwood from Washington and Lee University.